refactor the rbac integ tests to reduce test time
1) Puts all tests in the same package (main_test.go) to reuse the same
Istio cluster and avoid creating/deleting the Istio cluster multiple
times. Each test case will deploy the RBAC policy in their own namespace.

2) Do not wait for 60 seconds in each test case. The RunRBACTest()
function will just retry for 10 times. So if there is any delay in
policy propagation, it should be covered by the retry already, so it
doesn't make too much sense to wait for another 60 seconds.
