+++ tests/integration/security/rbac/testdata/clusterrbacconfig.yaml
