+++ tests/integration/security/rbac/testdata/v1-policy-group.yaml.tmpl
