+++ tests/integration/security/rbac/testdata/v1-policy-grpc.yaml.tmpl
