+++ tests/integration/security/rbac/testdata/v1-policy-path.yaml.tmpl
