+++ tests/integration/security/rbac/testdata/v2-policy-basic.yaml.tmpl
