+++ tests/integration/security/rbac/testdata/v2-policy-extended.yaml.tmpl
