+++ tests/integration/security/rbac/testdata/v2-policy-group.yaml.tmpl
