+++ tests/integration/security/rbac/testdata/v2-policy-grpc.yaml.tmpl
