--- tests/integration/security/rbac/v1/grpc/testdata/istio-clusterrbacconfig.yaml.tmpl
@@ -1,10 +1,0 @@
-# Enable istio RBAC
-apiVersion: "rbac.istio.io/v1alpha1"
-kind: ClusterRbacConfig
-metadata:
-  name: default
-spec:
-  mode: 'ON_WITH_INCLUSION'
-  inclusion:
-    namespaces: ["{{ .Namespace }}"]
----
