--- tests/integration/security/rbac/v2/basic/grpc_test.go
@@ -1,103 +1,0 @@
-// Copyright 2019 Istio Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package basic
-
-import (
-	"testing"
-	"time"
-
-	"istio.io/istio/tests/integration/security/util"
-
-	"istio.io/istio/pkg/test/echo/common/scheme"
-	"istio.io/istio/pkg/test/framework"
-	"istio.io/istio/pkg/test/framework/components/echo"
-	"istio.io/istio/pkg/test/framework/components/echo/echoboot"
-	"istio.io/istio/pkg/test/framework/components/environment"
-	"istio.io/istio/pkg/test/framework/components/namespace"
-	"istio.io/istio/pkg/test/util/file"
-	"istio.io/istio/pkg/test/util/tmpl"
-	rbacUtil "istio.io/istio/tests/integration/security/rbac/util"
-	"istio.io/istio/tests/integration/security/util/connection"
-)
-
-const (
-	grpcRbacV2RulesTmpl = "testdata/istio-rbac-grpc-rules.yaml.tmpl"
-)
-
-func TestRBACV2GRPC(t *testing.T) {
-	framework.NewTest(t).
-		RequiresEnvironment(environment.Kube).
-		Run(func(ctx framework.TestContext) {
-			ns := namespace.NewOrFail(t, ctx, "rbacv2-grpc-test", true)
-			var a, b, c, d echo.Instance
-			echoboot.NewBuilderOrFail(t, ctx).
-				With(&a, util.EchoConfig("a", ns, false, nil, g, p)).
-				With(&b, util.EchoConfig("b", ns, false, nil, g, p)).
-				With(&c, util.EchoConfig("c", ns, false, nil, g, p)).
-				With(&d, util.EchoConfig("d", ns, false, nil, g, p)).
-				BuildOrFail(t)
-
-			cases := []rbacUtil.TestCase{
-				{
-					Request: connection.Checker{
-						From: b,
-						Options: echo.CallOptions{
-							Target:   a,
-							PortName: "grpc",
-							Scheme:   scheme.GRPC,
-						},
-					},
-					ExpectAllowed: isMtlsEnabled,
-				},
-				{
-					Request: connection.Checker{
-						From: c,
-						Options: echo.CallOptions{
-							Target:   a,
-							PortName: "grpc",
-							Scheme:   scheme.GRPC,
-						},
-					},
-					ExpectAllowed: false,
-				},
-				{
-					Request: connection.Checker{
-						From: d,
-						Options: echo.CallOptions{
-							Target:   a,
-							PortName: "grpc",
-							Scheme:   scheme.GRPC,
-						},
-					},
-					ExpectAllowed: isMtlsEnabled,
-				},
-			}
-
-			namespaceTmpl := map[string]string{
-				"Namespace": ns.Name(),
-			}
-
-			policies := tmpl.EvaluateAllOrFail(t, namespaceTmpl,
-				file.AsStringOrFail(t, rbacClusterConfigTmpl),
-				file.AsStringOrFail(t, grpcRbacV2RulesTmpl))
-			g.ApplyConfigOrFail(t, ns, policies...)
-			defer g.DeleteConfigOrFail(t, ns, policies...)
-
-			// Sleep 60 seconds for the policy to take effect.
-			// TODO: Check to make sure policies have been created instead.
-			time.Sleep(60 * time.Second)
-			rbacUtil.RunRBACTest(t, cases)
-		})
-}
